Squashed unsigned vs. signed comparison warning.

+++ src/g3d/Command.cxx
@@ -72,7 +72,7 @@ const QStringList Command::getArgumentNames() const
 QString Command::getSyntax() const
 {
   QString line = _name;
-  for (size_t i = 0; i < _argNames.size(); ++i) {
+  for (int i = 0; i < _argNames.size(); ++i) {
     line += " <" + _argNames[i] + ">";
   }
   return line;
